TODO:
	mousewheel movement was disabled in sdl_input.c
	r_mode -2 was disabled in sdl_glimp.c
	joystick was disabled

Conflicts:
	src/sdl/sdl_gamma.c
	src/sdl/sdl_glimp.c
	src/sdl/sdl_input.c
	src/sys/sys_platform.c

+++ src/qcommon/md4.c
@@ -33,7 +33,7 @@
 
 /* GLOBAL.H - RSAREF types and constants */
 
-#include "../qcommon/q_shared.h"
+#include "q_shared.h"
 
 #ifdef _WIN32
 #ifndef __GNUC__
@@ -47,8 +47,8 @@ typedef unsigned char *POINTER;
 /* UINT2 defines a two byte word */
 typedef unsigned short int UINT2;
 
-/* UINT4 defines a four byte word */
-typedef unsigned long int UINT4;
+/* UINT4 defines a four byte word - was unsigned long int which did not work for 64 bit */
+typedef unsigned int UINT4;
 
 
 /* MD4.H - header file for MD4C.C */
